[sprint_5-game_start_screen] Вынос логики в компонент GamePageLauncher
