+++ packages/client/src/pages/game/game.tsx
@@ -4,48 +4,15 @@ import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import styles from './styles.module.scss';
-
-const descriptionPart = {
-    name: 'Destroy This Bubbles',
-    rules: [
-        'Щёлкните по пузырям, чтобы уничтожить их',
-        'Чем больше пузырей уничтожаете, тем больше очков вы получаете',
-        'За пропущенные пузыри вы теряете очки',
-    ],
-};
-
-const lastTryPart = {
-    lastGamesScore: [
-        {
-            date: '12.11.2025 12:00',
-            score: 100,
-            accuracy: 100,
-        },
-        {
-            date: '11.11.2025 12:00',
-            score: 76,
-            accuracy: 95,
-        },
-        {
-            date: '10.11.2025 12:00',
-            score: 50,
-            accuracy: 80,
-        },
-        {
-            date: '09.11.2025 12:00',
-            score: 25,
-            accuracy: 70,
-        },
-    ],
-};
+import { GamePageLauncher } from './launcher';
 
 const GamePage = () => {
     const navigate = useNavigate();
 
     const [countdown, setCountdown] = useState(3);
     const [gameState, setGameState] = useState<
-        'idle' | 'countdown' | 'playing'
-    >('idle');
+        'launcher' | 'countdown' | 'playing'
+    >('launcher');
 
     useEffect(() => {
         if (gameState === 'countdown') {
@@ -68,44 +35,8 @@ const GamePage = () => {
 
     return (
         <Page className={styles.container}>
-            {gameState === 'idle' && (
-                <>
-                    <div className={styles.descriptionPart}>
-                        <ul className={styles.list}>
-                            {descriptionPart.rules.map((rule) => (
-                                <li key={rule}>{rule}</li>
-                            ))}
-                        </ul>
-                        <Button
-                            variant="outlined"
-                            onClick={() => navigate('/menu')}
-                        >
-                            Назад к меню
-                        </Button>
-                    </div>
-                    <div className={styles.startPart}>
-                        <div
-                            className={styles.gameTitle}
-                            onClick={handleGameStart}
-                        >
-                            {descriptionPart.name}
-                        </div>
-                    </div>
-                    <div className={styles.lastTryPart}>
-                        <h2 className={styles.title}>Предыдущие игры</h2>
-                        {lastTryPart.lastGamesScore.map((game) => (
-                            <div key={game.date}>
-                                <p className={styles.date}>{game.date}</p>
-                                <p className={styles.info}>
-                                    Очки: {game.score}
-                                </p>
-                                <p className={styles.info}>
-                                    Точность: {game.accuracy}%
-                                </p>
-                            </div>
-                        ))}
-                    </div>
-                </>
+            {gameState === 'launcher' && (
+                <GamePageLauncher handleGameStart={handleGameStart} />
             )}
 
             {gameState === 'countdown' && <div>{countdown}</div>}
