+++ packages/client/src/pages/game/launcher/index.ts
@@ -1,0 +1,3 @@
+import GamePageLauncher from './launcher';
+
+export { GamePageLauncher };
