+++ packages/client/src/pages/game/launcher/styles.module.scss
@@ -1,0 +1,139 @@
+.container {
+    width: 100vw;
+    height: 100vh;
+
+    display: flex;
+
+    .descriptionPart {
+        display: flex;
+        flex-direction: column;
+        gap: 1rem;
+        justify-content: start;
+
+        padding: 1rem;
+        width: 25%;
+
+        color: #898484;
+
+
+        p {
+            font-style: italic;
+        }
+
+        .list {
+            margin-left: 1rem;
+        }
+    }
+
+    .startPart {
+        display: flex;
+        flex-direction: column;
+        gap: 1rem;
+        align-items: center;
+        justify-content: center;
+
+        padding: 1rem;
+        width: 50%;
+    }
+
+    .lastTryPart {
+        display: flex;
+        flex-direction: column;
+        gap: 1rem;
+        align-items: start;
+        justify-content: end;
+
+        padding: 1rem;
+        width: 25%;
+
+        color: #898484;
+
+        .title {
+            font-size: 1.5rem;
+        }
+
+        .date {
+            font-weight: bold;
+            font-style: italic;
+        }
+
+        .info { 
+            font-style: italic;
+        }
+    }
+
+    .gameTitle {
+        position: relative;
+        padding: 0.5rem 1rem;
+        cursor: crosshair;
+        font-size: 3rem;
+        font-weight: bold;
+        background: linear-gradient(45deg, #00ff00, #32cd32, #228b22, #006400);
+        background-size: 400% 400%;
+        -webkit-background-clip: text;
+        background-clip: text;
+        -webkit-text-fill-color: transparent;
+        text-shadow: 
+            0 0 10px #00ff0080,
+            0 0 20px #00ff0060,
+            0 0 30px #00ff0040,
+            0 0 40px #32cd3240,
+            0 0 50px #228b2230;
+        animation: fireFlicker 2s ease-in-out infinite alternate,
+                   backgroundShift 4s ease-in-out infinite;
+
+        &::before {
+            content: '';
+            position: absolute;
+            top: 0;
+            left: 0;
+            right: 0;
+            bottom: 0;
+            border: 2px solid #00ff00;
+            border-radius: 10px;
+            box-shadow: 0 0 5px #00ff00;
+        }
+    }
+}
+
+@keyframes fireFlicker {
+    0% {
+        text-shadow: 
+            0 0 10px #00ff0080,
+            0 0 20px #00ff0060,
+            0 0 30px #00ff0040,
+            0 0 40px #32cd3240,
+            0 0 50px #228b2230;
+        transform: scale(1);
+    }
+    50% {
+        text-shadow: 
+            0 0 15px #00ff00a0,
+            0 0 25px #00ff0080,
+            0 0 35px #00ff0060,
+            0 0 45px #32cd3260,
+            0 0 55px #228b2250;
+        transform: scale(1.02);
+    }
+    100% {
+        text-shadow: 
+            0 0 8px #00ff0070,
+            0 0 18px #00ff0050,
+            0 0 28px #00ff0030,
+            0 0 38px #32cd3230,
+            0 0 48px #228b2220;
+        transform: scale(0.98);
+    }
+}
+
+@keyframes backgroundShift {
+    0% {
+        background-position: 0% 50%;
+    }
+    50% {
+        background-position: 100% 50%;
+    }
+    100% {
+        background-position: 0% 50%;
+    }
+}
